chore: PhotographerSection 파일 분리
